Some changes to native conversion.
A type family 'NativeRowsG', and corresponding user-visible
'NativeRows', replaces the second type parameter in the *Native
typeclasses.  This exposes native conversion directly at the type level.

In addition, the 'ToNative' class was removed and 'ToNativeExact' was
renamed to it, because @toNative = toNativeExact . restrict@.  This
eliminates some code duplication and a bit of conceptual complexity in
understanding native conversion.